oef connection async refactoring and tests coverage
