+++ packages/hashes.csv
@@ -23,7 +23,7 @@ fetchai/connections/http_client,QmXQrA6gA4hMEMkMQsEp1MQwDEqRw5BnnqR4gCrP5xqVD2
 fetchai/connections/http_server,QmPMSyX1iaWM7mWqFtW8LnSyR9r88RzYbGtyYmopT6tshC
 fetchai/connections/ledger,QmezMgaJkk9wbQ4nzURERnNJdrzkQyvV5PiieH6uGbVzc3
 fetchai/connections/local,QmVcTEJxGbWbtXi2fLN5eJA6XuEAneaNd83UJPugrtb9xU
-fetchai/connections/oef,QmfX6fF2CqruwVc46Tqogb7SyyLEQa2t5J6SpN5wkj2tQw
+fetchai/connections/oef,QmY19byxGM3mADRSW8YVmrqTn7pPYuBZzkB3qU6pZmM8yj
 fetchai/connections/p2p_client,QmbwCDuAB1eq6JikqeAAqpqjVhxevGNeWCLqRD67Uvqiaz
 fetchai/connections/p2p_libp2p,QmdFoDC26e94ACZB2nVLTyoSMDwKGuyupB3WJhfZ2Mi3Bk
 fetchai/connections/p2p_libp2p_client,QmVhsh863k3ws4HeDpkZm7GQkrW3aMREu5sLkHATmwCddC
