+++ tests/test_packages/test_connections/test_oef/test_communication.py
@@ -16,6 +16,8 @@
 #   limitations under the License.
 #
 # ------------------------------------------------------------------------------
+
+
 """This test module contains the tests for the OEF communication using an OEF."""
 
 import asyncio
@@ -25,6 +27,7 @@ import time
 import unittest
 from typing import Dict, cast
 from unittest import mock
+from unittest.mock import patch
 
 from oef.messages import OEFErrorOperation
 from oef.query import ConstraintExpr
@@ -81,7 +84,8 @@ class OefSearchDialogues(BaseOefSearchDialogues):
 
     @staticmethod
     def role_from_first_message(message: Message) -> BaseDialogue.Role:
-        """Infer the role of the agent from an incoming/outgoing first message
+        """
+        Infer the role of the agent from an incoming/outgoing first message.
 
         :param message: an incoming/outgoing first message
         :return: The role of the agent
@@ -1089,7 +1093,7 @@ class TestSendWithOEF(UseOef):
         await oef_connection.connect()
 
         with unittest.mock.patch.object(
-            oef_connection.in_queue, "get", side_effect=Exception
+            oef_connection.channel._in_queue, "get", side_effect=Exception
         ):
             result = await oef_connection.receive()
             assert result is None
@@ -1124,20 +1128,48 @@ async def test_cannot_connect_to_oef():
         oef_addr="127.0.0.1",
         oef_port=61234,  # use addr instead of hostname to avoid name resolution
     )
-    oef_connection.loop = asyncio.get_event_loop()
+
     patch = unittest.mock.patch.object(
         packages.fetchai.connections.oef.connection.logger, "warning"
     )
     mocked_logger_warning = patch.start()
 
-    async def try_to_connect():
-        await oef_connection.connect()
-
-    task = asyncio.ensure_future(try_to_connect(), loop=asyncio.get_event_loop())
+    task = asyncio.ensure_future(
+        oef_connection.connect(), loop=asyncio.get_event_loop()
+    )
     await asyncio.sleep(1.0)
     mocked_logger_warning.assert_called_with(
         "Cannot connect to OEFChannel. Retrying in 5 seconds..."
     )
     await cancel_and_wait(task)
-    oef_connection.channel.disconnect()
-    oef_connection.disconnect()
+    await oef_connection.disconnect()
+
+
+@pytest.mark.asyncio
+@pytest.mark.skipif(
+    sys.version_info < (3, 7), reason="Python version < 3.7 not supported by the OEF."
+)
+async def test_methods_with_logging_only():
+    """Test the case when we can't connect to the OEF."""
+    oef_connection = _make_oef_connection(
+        address=FETCHAI_ADDRESS_ONE,
+        oef_addr="127.0.0.1",
+        oef_port=61234,  # use addr instead of hostname to avoid name resolution
+    )
+
+    with patch.object(oef_connection.channel, "_oef_agent_connect", return_value=True):
+        await oef_connection.connect()
+
+    oef_connection.channel.on_cfp(
+        msg_id=1, dialogue_id=1, origin="some", target=1, query=b""
+    )
+    oef_connection.channel.on_decline(msg_id=1, dialogue_id=1, origin="some", target=1)
+    oef_connection.channel.on_propose(
+        msg_id=1, dialogue_id=1, origin="some", target=1, proposals=b""
+    )
+    oef_connection.channel.on_accept(msg_id=1, dialogue_id=1, origin="some", target=1)
+
+    try:
+        await oef_connection.disconnect()
+    except Exception:  # nosec
+        pass
