+++ tests/test_packages/test_connections/test_oef/test_oef_serializer.py
@@ -16,7 +16,6 @@
 #   limitations under the License.
 #
 # ------------------------------------------------------------------------------
-
 """This test module contains the tests for the OEF serializer."""
 
 from aea.helpers.search.models import (
@@ -25,16 +24,26 @@ from aea.helpers.search.models import (
     ConstraintType,
     DataModel,
     Description,
+    Location,
     Query,
 )
 
+from packages.fetchai.connections.oef.object_translator import OEFObjectTranslator
 from packages.fetchai.protocols.oef_search.message import OefSearchMessage
 
 
 def test_oef_serialization_description():
     """Testing the serialization of the OEF."""
-    foo_datamodel = DataModel("foo", [Attribute("bar", int, True, "A bar attribute.")])
-    desc = Description({"bar": 1}, data_model=foo_datamodel)
+    foo_datamodel = DataModel(
+        "foo",
+        [
+            Attribute("bar", int, True, "A bar attribute."),
+            Attribute("location", Location, True, "A location attribute."),
+        ],
+    )
+    desc = Description(
+        {"bar": 1, "location": Location(10.0, 10.0)}, data_model=foo_datamodel
+    )
     msg = OefSearchMessage(
         performative=OefSearchMessage.Performative.REGISTER_SERVICE,
         dialogue_reference=(str(1), ""),
@@ -44,6 +53,23 @@ def test_oef_serialization_description():
     assert len(msg_bytes) > 0
     recovered_msg = OefSearchMessage.serializer.decode(msg_bytes)
     assert recovered_msg == msg
+
+
+def test_oef_object_transator():
+    """Test oef description and description tranlations."""
+    foo_datamodel = DataModel(
+        "foo",
+        [
+            Attribute("bar", int, True, "A bar attribute."),
+            Attribute("location", Location, True, "A location attribute."),
+        ],
+    )
+    desc = Description(
+        {"bar": 1, "location": Location(10.0, 10.0)}, data_model=foo_datamodel
+    )
+    oef_desc = OEFObjectTranslator.to_oef_description(desc)
+    new_desc = OEFObjectTranslator.from_oef_description(oef_desc)
+    assert desc.values["location"] == new_desc.values["location"]
 
 
 def test_oef_serialization_query():
